splits quadrants into two components, fix arrow opacity
